tdlib: Rename action module to chat_action
